requirements.txt
streamlit
pandas
plotly
pillow
openpyxl